requirements.txt
# Web framework
fastapi>=0.111.0

# ASGI server
uvicorn[standard]>=0.29.0

# Automação com navegador
selenium>=4.21.0
webdriver-manager>=4.0.1

# Planilhas e dados
pandas>=2.2.2
openpyxl>=3.1.2

# Autenticação
PyJWT>=2.8.0

# Configuração via Pydantic v2
pydantic-settings>=2.0.3

# Compatibilidade com Pydantic v2
pydantic>=2.6.4

pydantic[email]
datetime
httpx
docker
